requirements.txt
fastapi==0.95.2
uvicorn==0.20.0
pydantic==1.10.7
gspread==5.11.1
google-auth==2.22.0
google-cloud-vision==3.4.1
requests==2.31.0
Pillow==9.5.0
python-dotenv==1.0.0

